test : 로또 개수 확인
test코드 파일 분리, 숫자 범위 및 개수 확인
